refactor: bucket is its own file now
