+++ src/model.rs
@@ -9,6 +9,7 @@ pub use hashing::*;
 pub mod chat;
 pub mod server;
 pub mod channel;
+pub mod bucket;
 pub mod error;
 pub mod log;
 pub mod message;
