+++ src/model/channel.rs
@@ -1,8 +1,6 @@
-use chrono::{DateTime, NaiveDate, Utc};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::model::message::Message;
 
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -24,39 +22,5 @@ impl Channel
             id: Uuid::now_v7().to_string(),
             name: name_sanitized,
         }
-    }
-}
-
-
-//doubt i need this in model
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Bucket
-{
-    pub id: String,
-    pub chat: Channel,
-    pub date: NaiveDate,
-    pub messages: Vec<Message>,
-}
-
-impl Bucket
-{
-    #[must_use]
-    pub fn new(chat: &Channel, date: &DateTime<Utc>) -> Self
-    {
-        Self
-        {
-            id: Uuid::now_v7().to_string(),
-            chat: chat.clone(),
-            date: date.date_naive(),   
-            messages: Vec::new(),
-        }
-    }
-}
-
-impl Bucket
-{
-    pub fn add_message(&mut self, message: Message)
-    {
-        self.messages.push(message);
     }
 }
